+++ .storybook/main.js
@@ -8,6 +8,7 @@ module.exports = {
   addons: [
     '@storybook/addon-links',
     '@storybook/addon-essentials',
+    '@storybook/addon-a11y',
     'storybook-addon-performance',
   ],
   typescript: {
@@ -28,7 +29,7 @@ module.exports = {
     // these options will include styled-system props
     reactDocgenTypescriptOptions: {
       skipPropsWithoutDoc: false,
-      // shouldExtractLiteralValuesFromEnum: true,
+      shouldExtractLiteralValuesFromEnum: true,
       // // don't include node_module props as you'll cause the machine to run out of memory on our repo
       propFilter: (prop) => {
         // document props from styled-system