+++ packages/button/src/GhostButton.tsx
@@ -4,21 +4,21 @@ import { BoxStyleProps } from '@realsystem/box';
 import { merge } from '@realsystem/utils';
 
 import { baseStyles, BoxAsButton } from './styles';
-import { ButtonStates, InternalButtonProps } from './types';
+import { ButtonIntents, ButtonStates, InternalButtonProps } from './types';
 
 type ButtonStyles = Record<ButtonStates, BoxStyleProps>;
 
 const defaultStyles: ButtonStyles = {
   default: merge(baseStyles.default, {
     color: 'color-text-brand',
-    backgroundColor: 'color-background-weak',
+    backgroundColor: 'color-background',
     _hover: {
       color: 'color-text-brand-strong',
       backgroundColor: 'color-background-brand-weakest',
     },
     _active: {
-      color: 'color-text-brand-strong',
-      backgroundColor: 'color-background-brand-weakest',
+      color: 'color-text-brand-stronger',
+      backgroundColor: 'color-background-brand-weaker',
     },
   }),
   loading: merge(baseStyles.loading, {
@@ -31,16 +31,72 @@ const defaultStyles: ButtonStyles = {
   }),
 };
 
-const GhostButton = forwardRef<
-  HTMLButtonElement,
-  Omit<InternalButtonProps, 'intent'>
->(({ children, buttonState, ...restProps }, ref): React.ReactElement => {
-  return (
-    <BoxAsButton {...defaultStyles[buttonState]} {...restProps} ref={ref}>
-      {children}
-    </BoxAsButton>
-  );
-});
+const dangerStyles: ButtonStyles = {
+  default: merge(baseStyles.default, {
+    color: 'color-text-danger',
+    backgroundColor: 'color-background',
+    _hover: {
+      color: 'color-text-danger-strong',
+      backgroundColor: 'color-background-danger-weakest',
+    },
+    _active: {
+      color: 'color-text-danger-stronger',
+      backgroundColor: 'color-background-danger-weaker',
+    },
+  }),
+  loading: merge(baseStyles.loading, {
+    color: 'color-text-danger-weaker',
+    backgroundColor: 'color-background-danger-weakest',
+  }),
+  disabled: merge(baseStyles.disabled, {
+    color: 'color-text-danger-weaker',
+    backgroundColor: 'color-background-danger-weakest',
+  }),
+};
+
+const neutralStyles: ButtonStyles = {
+  default: merge(baseStyles.default, {
+    color: 'color-text-neutral',
+    backgroundColor: 'color-background',
+    _hover: {
+      color: 'color-text-neutral-strong',
+      backgroundColor: 'color-background-neutral-weakest',
+    },
+    _active: {
+      color: 'color-text-neutral-stronger',
+      backgroundColor: 'color-background-neutral-weaker',
+    },
+  }),
+  loading: merge(baseStyles.loading, {
+    color: 'color-text-neutral-weaker',
+    backgroundColor: 'color-background-neutral-weakest',
+  }),
+  disabled: merge(baseStyles.disabled, {
+    color: 'color-text-neutral-weaker',
+    backgroundColor: 'color-background-neutral-weakest',
+  }),
+};
+
+const STYLE_MAP: {
+  [key in ButtonIntents]: ButtonStyles;
+} = {
+  default: defaultStyles,
+  danger: dangerStyles,
+  neutral: neutralStyles,
+};
+
+const GhostButton = forwardRef<HTMLButtonElement, InternalButtonProps>(
+  (
+    { children, buttonState, intent = 'default', ...restProps },
+    ref
+  ): React.ReactElement => {
+    return (
+      <BoxAsButton {...STYLE_MAP[intent][buttonState]} {...restProps} ref={ref}>
+        {children}
+      </BoxAsButton>
+    );
+  }
+);
 
 GhostButton.displayName = 'GhostButton';