+++ packages/icon/tools/build.js
@@ -1,0 +1,3 @@
+const build = require('../../../tools/rollup');
+
+build(require('../package.json'));