+++ packages/theme/src/index.ts
@@ -3,12 +3,13 @@
  */
 export * from './GlobalStyles';
 export * from './makeTheme';
+export { normalize as normalizeCSS } from './normalize';
 export * from './ThemeProvider';
 export * from './themes';
 export * from './tokens/makeColorRange';
 /*
  * Token utilities
  */
-export type { ThemeTokens } from './getToken';
+export type { ThemeScales, ThemeTokens } from './getToken';
 export { getToken } from './getToken';
 export type { DefaultTheme } from '@realsystem/styling';