+++ packages/theme/src/tokens/default/palette.ts
@@ -3,32 +3,29 @@ import { PaletteColors } from '../types';
 export type DefaultPaletteValues =
   | '#635bff'
   | '#0a2540'
-  | '#2e2e2e'
+  | '#333333'
   | '#198038'
   | '#f1c21b'
   | '#da1e28'
   | '#0043ce'
   | '#161616'
   | '#ffffff'
-  | '#f9f9f9'
-  | '#babbc7'
   | '#121c2d';
 
 export type DefaultPalette = { [key in PaletteColors]: DefaultPaletteValues };
 
 const palette: DefaultPalette = {
   brand: '#635bff',
-  inverse: '#f9f9f9',
   primary: '#0a2540',
-  secondary: '#2e2e2e',
+  secondary: '#121c2d',
   success: '#198038',
   warning: '#f1c21b',
   danger: '#da1e28',
   info: '#0043ce',
   strong: '#161616',
-  medium: '#121c2d',
+  medium: '#333333',
   weak: '#ffffff',
-  neutral: '#babbc7',
+  neutral: '#333333',
 };
 
 export default palette;