+++ tools/utils/makeEsbuildConfig.js
@@ -1,7 +1,7 @@
 const { getExternalDeps } = require('./getExternalDeps');
 const { ENV, isProduction } = require('./env');
 const { logger } = require('./logger');
-const { commandSync } = require('execa');
+const { command } = require('execa');
 
 // ESbuild config
 const baseEsbuildConfig = {
@@ -24,9 +24,10 @@ const shouldWatch = (pkg = {}) =>
   !isProduction
     ? {
         // eslint-disable-next-line no-unused-vars
-        onRebuild(err, _result) {
+        async onRebuild(err, _result) {
           if (err) logger.error(err);
-          commandSync('yarn types');
+          console.log(process.cwd());
+          // await command('yarn types');
           logger.blue(`Rebundled ${pkg.name}.`);
         },
       }