+++ packages/core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @realsystem/core
 
+## 1.7.5
+
+### Patch Changes
+
+- [`0d999f9`](https://github.com/bigwoof91/realsystem/commit/0d999f97098a4daa4cc69322fee33ea385a97078) - adding icon, fixing typography colors, restyling button with icon integration
+
+- Updated dependencies [[`0d999f9`](https://github.com/bigwoof91/realsystem/commit/0d999f97098a4daa4cc69322fee33ea385a97078)]:
+  - @realsystem/box@8.0.5
+  - @realsystem/button@8.0.5
+  - @realsystem/styling@1.6.5
+  - @realsystem/theme@2.0.5
+  - @realsystem/typography@0.0.2
+  - @realsystem/utils@1.5.5
+
 ## 1.7.4
 
 ### Patch Changes